requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
motor==3.7.1
reportlab==4.3.1
pandas==2.2.2
matplotlib==3.9.0
numpy==2.2.6
pymongo==4.14.1
